+++ include/openvswitch/automake.mk
@@ -53,7 +53,6 @@ openvswitchinclude_HEADERS = \
 
 if P4OVS
 openvswitchinclude_HEADERS += \
-	include/openvswitch/ovs-p4rt.h \
 	include/openvswitch/p4ovs.h
 endif
